Separando el codigo para poder mantenerlo mas fácil
