+++ src/functions/index.ts
@@ -1,0 +1,7 @@
+export * from "./getAnimeInfo"
+export * from "./getComing"
+export * from "./getLatest"
+export * from "./getOnAir"
+export * from "./searchAnime"
+export * from "./searchAnimeBySpecificURL"
+export * from "./searchAnimesByFilter"
