+++ src/functions/searchAnime.ts
@@ -1,0 +1,19 @@
+import cloudscraper from "cloudscraper";
+import { CloudscraperOptions } from "../constants";
+import { scrapSearchAnimeData } from "../utils/scrapAnimeData";
+import { load } from "cheerio";
+import { SearchAnimeData } from "../types";
+
+export async function searchAnime(query: string): Promise<SearchAnimeData[]> {
+	try {
+
+		CloudscraperOptions.uri = 'https://www3.animeflv.net/browse?q=' + query.toLowerCase().replace(/\s+/g, "+");
+
+		const searchData = (await cloudscraper(CloudscraperOptions)) as string;
+		const $ = load(searchData);
+
+		return scrapSearchAnimeData($);
+	} catch {
+		return [];
+	}
+}
